requirements.txt
streamlit>=1.46.0
openai>=1.91.0
pandas>=2.3.0
pydantic>=2.11.7
python-dotenv>=1.1.1 